+++ src/graphtea/extensions/reports/topological/GutmanIndex.java
@@ -37,7 +37,7 @@ public class GutmanIndex implements GraphReportExtension<Integer> {
                     	int degreeOfU = g.getDegree(g.getVertex(u));
                     	int degreeOfV = g.getDegree(g.getVertex(v));
 
-                    	sum += (degreeOfU * degreeOfV) * UVdist;
+                    	sum += (int) ((degreeOfU * degreeOfV) * UVdist);
                     }
                 }
             }
