+++ src/graphtea/extensions/reports/topological/MWienerIndex.java
@@ -33,7 +33,7 @@ public class MWienerIndex implements GraphReportExtension<Integer> {
                 if(spt[v][u] < g.numOfVertices() + 1) {
                     double dist = spt[u][v];
                     if(dist > max) {
-                        sum *= dist;
+                        sum *= (int) dist;
                     }
                 }
             }
